create a separate test class for writing garbage code :)
