+++ src/protobuf-net.Test/NullWrappedValueTests.cs
@@ -11,7 +11,7 @@ using Xunit.Abstractions;
 
 namespace ProtoBuf.Test
 {
-    public class NullWrappedValueTests
+    public partial class NullWrappedValueTests
     {
         private readonly ITestOutputHelper _log;
         public NullWrappedValueTests(ITestOutputHelper log)
@@ -21,22 +21,6 @@ namespace ProtoBuf.Test
         }
         private void Log(string message)
             => _log?.WriteLine(message);
-
-        [Fact]
-        public void ExistingListsBehaviour()
-        {
-            using var ms = new MemoryStream(new byte[] { 0x08, 0x00, 0x08, 0x01, 0x08, 0x02 });
-            var clone = Serializer.Deserialize<Foo>(ms);
-            if (!ms.TryGetBuffer(out var buffer)) buffer = new ArraySegment<byte>(ms.ToArray());
-            Assert.Equal("08-00-08-01-08-02", BitConverter.ToString(buffer.Array, buffer.Offset, buffer.Count));
-        }
-
-        [DataContract]
-        public class Foo
-        {
-            [DataMember(Order = 1)]
-            public List<int?> Items { get; } = new List<int?>();
-        }
 
         [Theory]
         [InlineData(true)]
